linearAlgebra.jl

linearAlgebra.jl

This unit contains linear algebra functions useful in relation to the Riemannian geometry of the manifold of Symmetric Positive Definite (SPD) or Hermitian Positive Definite (HPD) matrices. In Julia those are Hermitian matrices, see typecasting matrices.

In general they take a matrix as input (some may take other arrays as input) and are divided in seven categories depending on what kind of functions thay are and what they give as output:

Category	Output

1. Matrix normalizations	matrix

2. Boolean functions of matrices	matrix

3. Scalar functions of matrices	scalar

4. Diagonal functions of matrices	diagonal matrix

5. Unitary functions of matrices	orthogonal/unitary matrix

6. Matrix function of matrices	matrix

7. Spectral decompositions of positive matrices	spectral function of input

8. Decompositions involving triangular matrices	triangular matrix

⋅

Matrix normalizations

Function	Description

det1	Normalize the determinant

tr1	Normalize the trace

normalizeCol!	Normalize one or more columns

⋅

PosDefManifold.det1 — Function.

det1!(P::ℍ)

Given a positive definite matrix $P$, return the best approximant to $P$ from the set of matrices in the special linear group, i.e., the closer matrix having det=1. See Bhatia and Jain (2014)[🎓].

$P$ must be flagged as Hermitian. See typecasting matrices, however a catch-all method is defined.

Examples

using LinearAlgebra, PosDefManifold
P=randP(5) # generate a random real positive definite matrix 5x5
Q=det1(P)
det(Q) # must be 1

See: typecasting matrices.

See also: tr1, det.

PosDefManifold.tr1 — Function.

tr1(P::ℍ)

Given a positive definite matrix P, return the trace-normalizedP` (trace=1).

$P$ must be flagged as Hermitian. See typecasting matrices, however a catch-all method is defined.

Examples

using LinearAlgebra, PosDefManifold
P=randP(5) # generate a random real positive definite matrix 5x5
Q=tr1(P)
tr(Q)  # must be 1

See: typecasting matrices.

See also: det1, tr.

PosDefManifold.normalizeCol! — Function.

(1) normalizeCol!(X::Matrix, j::Int)
(2) normalizeCol!(X::Matrix, j::Int, by::Number)
(3) normalizeCol!(X::Matrix, range::UnitRange)
(4) normalizeCol!(X::Matrix, range::UnitRange, by::Number)

Given a general matrix $X$ comprised of real or complex elements,

(1) normalize the $j^{th}$column

(2) divide the elements of the $j^{th}$ column by number $by$

(3) normalize the columns in $range$

(4) divide the elements of columns in $range$ by number $by$.

$by$ is a number of abstract supertype Number. It should be an integer, real or complex number.

$range$ is a UnitRange type.

No range check nor type check is performed. A catch-all method is defined.

Examples

using PosDefManifold
X=randn(10, 20)
normalizeCol!(X, 2)                 # (1) normalize columns 2
normalizeCol!(X, 2, 10.0)           # (2) divide columns 2 by 10.0
normalizeCol!(X, 2:4)               # (3) normalize columns 2 to 4
X=randn(ComplexF64, 10, 20)
normalizeCol!(X, 3)                 # (1) normalize columns 3
normalizeCol!(X, 3:6, (2.0 + 0.5im))# (4) divide columns 3 to 5 by (2.0 + 0.5im)

See also: colNorm, norm, randn

Boolean functions of matrices

Function	Description

ispos	Check whether the argument is comprised of all positive elements

PosDefManifold.ispos — Function.

(1) ispos(  λ::Vector; <tol::Real=minpos, rev::Bool=true,
            bell::Bool=true, msg::String="">)
(2) ispos(  Λ::Diagonal; <tol::Real=minpos, rev::Bool=true,
            bell::Bool=true, msg::String="">)

Return $true$ if all numbers in (1) real vector $λ$ or in (2) real diagonal matrix $Λ$ are not inferior to $tol$, otherwise return $false$. This may be used, for example, in spectral functions to check that all eigenvalues are positive. The default value for $tol$ is constant $minpos$ declared in the main module constants.

The following are <optional keyword arguments>:

If $rev=true$ the (1) elements in $λ$ or (2) the diagonal elements in $Λ$ will be chacked in reverse order.

This is done for allowing a very fast check when the elements are sorted where to start checking.

If the result is $false$:

if $bell=true$ a bell character will be printed. In most systems this will ring a bell on the computer.

if string $msg$ is provided, a warning will print $msg$ followed by:

"at position pos", where pos is the position where the first non-positive element has been found.

 ## Examples
 using PosDefManifold
 a=[1, 0, 2, 8];
 ispos(a, msg="non-positive element found")

 # it will print:
 # ┌ Warning: non-positive element found at position 2
 # └ @ [here julie will point to the line of code issuing the warning]

Scalar functions of matrices

Function	Description

colProd	Sum of products of the elements in two columns

sumOfSqr	Sum of squares of all elements or of specified columns

sumOfSqrDiag	Sum of squares of the diagonal elements

colNorm	Eucliden norm of a column

sumOfSqrTril	Sum of squares of the lower triangle elements up to a given underdiagonal

fidelity	(Quantum) Fidelity of two positive matrices

⋅

PosDefManifold.colProd — Function.

colProd(X::Matrix, j::Int, l::Int)

Given a general matrix $X$, comprised of real or complex elements, return the dot product of the $j^{th}$ and $l^{th}$ columns, defined as,

$\sum_{i=1}^{r} \big(x_{ij}^*x_{il}\big),$

where $r$ is the number of rows of $X$ and $^*$ the complex conjugate.

No range check nor type check is performed. A catch-all method is defined.

Arguments $j$ and $l$ must be positive integers in range 1:size(X, 2).

Examples

using PosDefManifold
X=randn(10, 20)
p=colProd(X, 1, 3)

PosDefManifold.sumOfSqr — Function.

(1) sumOfSqr(A::Array)
(2) sumOfSqr(X::Matrix, j::Int)
(3) sumOfSqr(X::Matrix, range::UnitRange)

Return

(1) the sum of square of the elements in an array $A$ of any dimensions.

(2) the sum of square of the $j^{th}$ column of a matrix $X$.

(3) the sum of square of the columns of $X$ in a given range.

No range check nor type check is performed. A catch-all method is defined.

Note that only (1) works for arrays of any dimensions and that if $A$ is a matrix (1) returns the square of the Frobenius norm: $\sum |a_{ij}|^2.$

Arguments

(1) (A):

$A$ is an array of any dimensions (e.g., a vector, matrix or tensor), real or complex.

(2) (X, j):

$X$ is a generic matrix, real or complex;

$j$ is a positive integer in range 1:size(X, 2).

(3) (X, range):

$X$ is a generic matrix, real or complex;;

$range$ is a UnitRange type.

Examples

using PosDefManifold
X=randn(10, 20)
sum²=sumOfSqr(X)        # (1) sum of squares of all elements
sum²=sumOfSqr(X, 1)     # (2) sum of squares of elements in column 1
sum²=sumOfSqr(X, 2:4)   # (3) sum of squares of elements in column 2 to 4

PosDefManifold.sumOfSqrDiag — Function.

(1) sumOfSqrDiag(X::Matrix)
(2) sumOfSqrDiag(D::Diagonal)

Return (1) the sum of squares of the diagonal elements in general matrix $X$ comprised of real or complex numbers. If $X$ is rectangular, the main diagonal is considered.

It also return (2) the sum of squares of real diagonal matrix $Λ$.

No range check nor type check is performed. A catch-all method is defined.

Examples

using LinearAlgebra, PosDefManifold
X=randn(10, 20)
sumDiag²=sumOfSqrDiag(X) # (1)
sumDiag²=sumOfSqrDiag(Diagonal(X)) # (2)

PosDefManifold.colNorm — Function.

colNorm(X::Matrix, j::Int)

Return the Euclidean norm of the $j^{th}$ column of general matrix $X$. No range check nor type check is performed. A catch-all method is defined.

Examples

using PosDefManifold
X=randn(10, 20)
normOfSecondColumn=colNorm(X, 2)

PosDefManifold.sumOfSqrTril — Function.

sumOfSqrTril(X::Matrix, k::Int=0)

Given a general matrix $X$, return the sum of squares of the elements in the lower triangle $X$ up to the $k^{th}$ underdiagonal.

$X$ may be rectangular. $k$ must be in range 1-size(X, 1):0.

See julia tril(M, k::Integer) function for numbering of diagonals. No range check nor type check is performed. A catch-all method is defined.

Examples

using PosDefManifold
A=[4. 3.; 2. 5.; 1. 2.]
#3×2 Array{Float64,2}:
# 4.0  3.0
# 2.0  5.0
# 1.0  2.0

s=sumOfSqrTril(A, -1)
# 9.0 = 1²+2²+2²

s=sumOfSqrTril(A, 0)
# 50.0 = 1²+2²+2²+4²+5²

PosDefManifold.fidelity — Function.

fidelity(P::ℍ, Q::ℍ)

Given two positive definte matrices $P$ and $Q$, return their fidelity:

$tr\big(P^{1/2}QP^{1/2}\big)^{1/2}.$

This is used in quantum physics and is related to the Wasserstein metric. See for example Bhatia, Jain and Lim (2019b)🎓.

$P$ and $Q$ must be flagged as Hermitian. See typecasting matrices, however a catch-all method is defined.

Examples

using PosDefManifold
P=randP(5);
Q=randP(5);
f=fidelity(P, Q)

See: typecasting matrices

Diagonal functions of matrices

Function	Description

fDiagonal	Elemen-wise functions of matrix diagonals

⋅

PosDefManifold.fDiagonal — Function.

fDiagonal(X::Matrix, func::Function, k::Int=0)

Applies function func element-wise to the elements of the $k^{th}$ diagonal of generic matrix $X$ (real or complex) of dimension r⋅c and return a diagonal matrix with these elements. Note that the dimension of the result depends on the size of $X$ and the chosen diagonal.

For example,

r ≠ c and $k$=0 (main diagonal), the result will be of dimension min(r,c)⋅min(r,c),

$X$ 3⋅4 and $k=-1$, the result will be 2⋅2,

$X$ 3⋅4 and $k=1$, the result will be 3⋅3, etc.

See julia tril(M, k::Integer) function for numbering of diagonals.

Arguments (X, func, k)

$X$ is a generic matrix of real or complex elements;

func is a function;

$k$ is the chosen diagonal (by default k=0, which is the main diagonal).

Nota Bene

The function func must support the func. syntax and therefore must be able to apply element-wise to the elements of the chosen diagonal (this includes anonymous functions). If X is complex, the function func must be able to support complex arguments.

A catch-all method is defined.

Examples

using PosDefManifold
P=randP(5) # use P=randP(ComplexF64, 5) for generating an Hermitian matrix
(Λ, U) = evd(P)         # Λ holds the eigenvalues of P, see evd
Δ=fDiagonal(Λ, log)     # diagonal matrix with the log of the eigenvalues
Δ=fDiagonal(Λ, x->x^2)  # using an anonymous function for the square of the eigenvalues

Unitary functions of matrices

Function	Description

mgs	Modified Gram-Schmidt orthogonalization

⋅

PosDefManifold.mgs — Function.

mgs(T::Matrix, numCol::Int=0)

Modified (stabilized) Gram-Schmidt orthogonalization of the columns of square or tall matrix $T$, which can be comprised of real or complex elements. The orthogonalized $T$ is returned by the function.

$T$ is not changed.

All columns are orthogonalized by default. If instead argument numCol is provided, then only the first numCol columns of $T$ are orthogonalized. In this case only the firt `numCol columns will be returned.

Examples

using LinearAlgebra, PosDefManifold
X=randn(10, 10);
U=mgs(X)        # result is 10⋅10
U=mgs(X, 3)     # result is 10⋅3
U'*U ≈ I ? println(" ⭐ ") : println(" ⛔ ")
# julia undertands also:
U'U ≈ I ? println(" ⭐ ") : println(" ⛔ ")

Matrix function of matrices

Function	Description

none for now	ipse lorem...

⋅

Spectral decompositions of positive matrices

Function	Description

evd	Eigenvalue-Eigenvector decomposition of a matrix in $UΛU'=P$ form

spectralFunctions	Mother function for creating spectral functions of eigenvalues

pow	Power of a positive matrix for any number of exponents in one pass

invsqrt	Principal square root inverse (whitening) of a positive matrix

sqr	Square of a positive matrix

powerIterations, powIter	Power method for estimating any number of eigenvectors and associated eigenvalues

choL	Lower triangula factor of Cholesky decomposition

⋅

PosDefManifold.evd — Function.

evd(S::ℍ)

Given a positive semi-definite matrix $S$, returns a 2-tuple $(Λ, U)$, where $U$ is the matrix holding in columns the eigenvectors and $Λ$ is the matrix holding the eigenvalues on the diagonal. This is the output of Julia eigen function in $UΛU'=S$ form.

As for the eigen function, the eigenvalues and associated eigenvectors are sorted by increasing values of eigenvalues.

$S$ must be flagged by julia as Hermitian. See typecasting matrices and typecasting matrices.

Examples

using PosDefManifold
A=randn(3, 3);
S=ℍ(A+A');
Λ, U=evd(S); # which is equivalent to (Λ, U)=evd(P)
(U*Λ*U') ≈ S ? println(" ⭐ ") : println(" ⛔ ")
# => UΛU'=S, UΛ=SU, ΛU'=U'S

PosDefManifold.spectralFunctions — Function.

spectralFunctions(P::ℍ, func)

This is the mother function for all spectral functions of eigenvalues implemented in this library, which are:

pow (power),

isqrt (inverse square root).

The function sqr (square) does not use it, as it can be obtained more efficiently by simple multiplication.

You can use this function if you need another spectral function of eigenvalues besides those and those already implemented in the standard package LinearAlgebra. In general, you won't call it directly.

The definition of spectral functions for a positive definite matrix $P$ is at it follows:

$f\big(P\big)=Uf\big(Λ\big)U',$

where $U$ is the matrix holding in columns the eigenvectors of $P$, $Λ$ is the matrix holding on diagonal its eigenvalues and $f$ is a function applying element-wise to the eigenvalues.

Arguments (P, func);

$P$ is a positive matrix.

`func is the function that will be applied on the eigenvalues

Nota Bene

The function func must support the func. syntax and therefore must be able to apply element-wise to the eigenvalues (those include anonymous functions).

Examples

using LinearAlgebra, PosDefManifold
n=5
P=randP(n) # P=randP(ComplexF64, 5) to generate an Hermitian complex matrix
noise=0.1;
Q=spectralFunctions(P, x->x+noise) # add white noise to the eigenvalues
tr(Q)-tr(P) ≈ noise*n ? println(" ⭐ ") : println(" ⛔ ")

See: typecasting matrices.

See also: evd.

PosDefManifold.pow — Function.

pow(P::ℍ, p)        # one argument
pow(P::ℍ, args...)  # several arguments

Given a positive definite matrix $P$, return the power $P^{r_1}, P^{r_2},...$ for any number of exponents $r_1, r_2,...$. It returns a tuple of as many elements as arguments passed after $P$.

Arguments (P, arg1, arg2,...)

$P$ is a positive matrix.

$arg1, arg2,...$ are real numbers.

Examples

using LinearAlgebra, PosDefManifold
P=randP(5);     # use P=randP(ComplexF64, 5) for generating an Hermitian matrix
Q=pow(P, 0.5);            # =>  QQ=P
Q, W=pow(P, 0.5, -0.5);
W*P*W ≈ I ? println(" ⭐ ") : println(" ⛔ ")
Q*Q ≈ P ? println(" ⭐ ") : println(" ⛔ ")
R, S=pow(P, 0.3, 0.7);
R*S ≈ P ? println(" ⭐ ") : println(" ⛔ ")

See: typecasting matrices.

See also: invsqrt.

PosDefManifold.invsqrt — Function.

invsqrt(P::ℍ)

Given a positive definite matrix $P$, compute the inverse of the principal square root $P^{-1/2}$.

Examples

using LinearAlgebra, PosDefManifold
P=randP(ComplexF64, 5);
Q=invsqrt(P);
Q*P*Q ≈ I ? println(" ⭐ ") : println(" ⛔ ")

See: typecasting matrices.

See also: pow.

PosDefManifold.sqr — Function.

sqr(P::ℍ)

Given a positive definite matrix $P$, compute its square $P^{2}$.

Examples

using PosDefManifold
P=randP(5);
P²=sqr(P);  # =>  P²=PP
sqrt(P²)≈ P ? println(" ⭐ ") : println(" ⛔ ")

See: typecasting matrices

PosDefManifold.powerIterations — Function.

powerIterations(S::ℍ, q;
        <evalues=false, tol=1e-9, maxiter=300, ⍰=false>)

alias: powIter

Compute the $q$ eigenvectors associated to the $q$ largest (real) eigenvalues of matrix $S$ using the power iterations + Gram-Schmidt orthogonalization as suggested by Strang.

$S$ must be flagged by julia as Hermitian. See typecasting matrices.

Nota Bene

Differently from the evd function, the eigenvectors and eigenvalues are sorted by decreasing order of eigenvalues.

Arguments (S, q; evalues=false, tol=1e-9, maxiter=300, ⍰=false):

$S$ is an Hermitian matrix (real, complex, positive definite or not).

$q$ is the number of eigenvectors to be found (thus the number of columns of the output).

The following are <optional keyword arguments>:

$tol$ is the tolerance for the convergence of the power method.

$maxiter$ is the maximum number of iterations allowed for the power method

if $⍰=true$, the convergence of all iterations will be printed.

if $evalues=true$, return the 4-tuple $(Λ, U, iterations, covergence)$

if $evalues=false$ return the 3-tuple $(U, iterations, covergence)$

Examples

using LinearAlgebra, PosDefManifold
S=randP(10);
# all eigenvectors
U, iterations, covergence=powIter(S, size(P, 2), ⍰=true)
# 3 eigenvectors and eigenvalues
Λ, U, iterations, covergence=powIter(S, 3, evalues=true);
U'*U≈ I ? println(" ⭐ ") : println(" ⛔ ")

See also: mgs

Decompositions involving triangular matrices

PosDefManifold.choL — Function.

choL(P::ℍ)
choL(P::Matrix)

Given a positive matrix $P$, return the Cholesky lower triangular factor $L$ such that $LL'=P$. To obtain $L'$ or both $L$ and $L'$, use instead julia function cholesky(P).

$P$ sould be flagged as Hermitian - see typecasting matrices - but a method for generic matrices is also provided.

On output, $L$ is of type LowerTriangular.

Examples

using PosDefManifold
P=randP(5);
L=choL(P);
L*L'≈ P ? println(" ⭐ ") : println(" ⛔ ")