riemannianGeometry.jl

riemannianGeometry.jl

This is the fundamental unit of PosDefManifold. It contains functions for manipulating points in the Riemannian manifold of Symmetric Positive Definite (SPD) or Hermitian Positive Definite (HPD) matrices. In Julia those are Hermitian matrices, see typecasting matrices.

The functions are divided in six categories:

Category	Output

1. Geodesic equations	interpolation, extrapolation,...

2. Distances	length of geodesics

3. Graphs and Laplacians	for spectral embedding, eigenmaps, system dynamics,...

4. Means	mid-points of geodesics, centers of mass of several points

5. Tangent Space operations	maps from the manifold to the tangent space and viceversa

6. Procrustes problems	for data matching, transfer learning,...

⋅

Geodesic equations

PosDefManifold.geodesic — Function.

geodesic(P::ℍ, Q::ℍ, a::Real, metric::Metric=Fisher)

Move along the geodesic from point $P$ to point $Q$ (two positive definite matrices) with arclegth $0<=a<=1$, using the specified metric, of type Metric::Enumerated type. By default de Fisher metric is adopted.

For all metrics,

with $a=0$ we stay at $P$,

with $a=1$ we move up to $Q$,

with $a=1/2$ we move to the mid-point of $P$ and $Q$ (mean).

Using the Fisher metric argument a can be any real number, for instance:

with $0<a<1$ we move toward $Q$ (attraction),

with $a>1$ we move over and beyond $Q$ (extrapolation),

with $a<0$ we move back away from Q (repulsion).

Note that if $Q=I$, the Fisher geodesic move is simply $P^a$ (no need to call this funtion then).

For the logdet zero and Jeffrey metric no closed form expression for the geodesic is available (to the best of authors' knowledge), so in this case the geodesic is found as the weighted mean [meanP(@ref)]. For the Von Neumann not even an expression for the mean is available, so in this case the geodesic is not provided and a warning is printed.

$P$ and $Q$ must be flagged by julia as Hermitian. See typecasting matrices.

Maths

For points $P$, $Q$ and arclength $a$, letting $b=1-a$, the geodesic equations for the supported metrics are:

Metric	geodesic equation

Euclidean	$bP + aQ$

invEuclidean	$\big(bP^{-1} + aQ^{-1}\big)^{-1}$

ChoEuclidean	$TT^*$, where $T=bL_P + aL_Q$

logEuclidean	$\text{exp}\big(b\text{log}(P) + a\text{log}(Q)\big)$

logCholesky	$TT^*$, where $T=S_P+a(S_Q-S_P)+D_P\hspace{2pt}\text{exp}\big(a(\text{log}D_Q-\text{log}D_P)\big)$

Fisher	$P^{1/2} \big(P^{-1/2} Q P^{-1/2}\big)^a P^{1/2}$

logdet0	uses weighted mean algorithm logdet0Mean

Jeffrey	uses weighted mean meanP

VonNeumann	N.A.

Wasserstein	$b^2P+a^2Q +ab\big[(PQ)^{1/2} +(QP)^{1/2}\big]$

legend: $L_X$, $S_X$ and $D_X$ are the Cholesky lower triangle of $X$, its strictly lower triangular part and diagonal part, respectively (hence, $S_X+D_X=L_X$, $L_XL_X^*=X$).

See also: meanP

Examples

using PosDefManifold
P=randP(10);
Q=randP(10);
# Wasserstein mean
M=geodesic(P, Q, 0.5, Wasserstein)
# extrapolate suing the Fisher metric
E=geodesic(P, Q, 2)

source

Distances

PosDefManifold.distanceSqr — Function.

(1) distanceSqr(P::ℍ, metric::Metric=Fisher
(2) distanceSqr(P::ℍ, Q::ℍ, metric::Metric=Fisher)

alias: distance²

(1) Return $δ^2(P, I)$, the square of the distance (or divergence) of positive definite matrix $P$ from the the identity matrix. See distance from the origin.

(2) Return $δ^2(P, Q)$, the square of the distance (or divergence) between two positive definite matrices $P$ and $Q$. See distance.

In both cases the distance function $δ$ is induced by the argument metric of type Metric::Enumerated type. By default, the Fisher metric is adopted.

The distance is always real, non-negative and equal to zero if and only if (1) $P=I$ or (2) $P=Q$.

$P$ in (1) and $P$, $Q$ in (2) must be flagged by julia as Hermitian. See typecasting matrices.

Maths

For point $P$ the squared distances from the identity for the supported metrics are:

Metric	Squared Distance from the identity

Euclidean	$∥P-I∥^2$

invEuclidean	$∥P^{-1}-I∥^2$

ChoEuclidean	$∥L_P-I∥^2$

logEuclidean	$∥\textrm{log}P∥^2$

logCholesky	$∥S_P∥^2+∥\textrm{log}D_P∥^2$

Fisher	$∥\textrm{log}P∥^2$

logdet0	$\textrm{logdet}\frac{1}{2}(P+I) - \frac{1}{2}\textrm{logdet}(P)$

Jeffrey	$\frac{1}{2}\textrm{tr}(P+P^{-1})-n$

VonNeumann	$\frac{1}{2}\textrm{tr}(P\textrm{log}P-\textrm{log}P)$

Wasserstein	$\textrm{tr}(P+I) -2\textrm{tr}(P^{1/2})$

For points $P$ and $Q$ their squared distances for the supported metrics are:

Metric	Squared Distance

Euclidean	$∥P-Q∥^2$

invEuclidean	$∥P^{-1}-Q^{-1}∥^2$

ChoEuclidean	$∥ L_P - L_Q ∥^2$

logEuclidean	$∥\textrm{log}P-\textrm{log}Q∥^2$

logCholesky	$∥S_P-S_Q∥^2+∥\textrm{log}D_P-\textrm{log}D_Q∥^2$

Fisher	$∥\textrm{log}(P^{-1/2}QP^{-1/2})∥^2$

logdet0	$\textrm{logdet}\frac{1}{2}(P+Q) - \frac{1}{2}\textrm{logdet}(PQ)$

Jeffrey	$\frac{1}{2}\textrm{tr}(Q^{-1}P+P^{-1}Q)-n$

VonNeumann	$\frac{1}{2}\textrm{tr}(P\textrm{log}P-P\textrm{log}Q+Q\textrm{log}Q-Q\textrm{log}P)$

Wasserstein	$\textrm{tr}(P+Q) -2\textrm{tr}(P^{1/2}QP^{1/2})^{1/2}$

legend: $L_X$, $S_X$ and $D_X$ are the Cholesky lower triangle of $X$, its strictly lower triangular part and diagonal part, respectively (hence, $S_X+D_X=L_X$, $L_XL_X^*=X$).

See also: distanceSqrMat

Examples (1)

using PosDefManifold
P=randP(10);
d=distanceSqr(P, Wasserstein)
e=distanceSqr(P) # uses the default metric (Fisher)
metric=Metric(Int(logdet0)) # or metric=logdet0
s=string(metric) # check what is the current metric
f=distance²(P, metric) #using the alias distance²

Examples (2)

using PosDefManifold
P=randP(10);
Q=randP(10);
d=distanceSqr(P, Q, Wasserstein)
e=distance²(P, Q, Jeffrey)

source

PosDefManifold.distance — Function.

(1) distance(P::ℍ, metric::Metric=Fisher
(2) distance(P::ℍ, Q::ℍ, metric::Metric=Fisher)

(1) Return $δ(P, I)$, the distance between positive definite matrix $P$ and the identity matrix.

(2) Return $δ(P, Q)$, the distance between positive definite matrices $P$ and $Q$.

This is the square root of distanceSqr and is invoked with the same syntax therein.

See also: distanceMatrix

source

Graphs and Laplacians

PosDefManifold.distanceSqrMat — Function.

distanceSqrMat(℘, metric::Metric=Fisher)

alias: distance²Mat

Given a 1d array ℘ of $k$ positive definite matrices ${P_1,...,P_k}$, create the $k⋅k$ real Hermitian matrix comprising elements $δ^2(P_i, P_j)\textrm{, for all }i≠j$.

This is the matrix of all squared inter-distances (zero on diagonal), using the specified metric, of type Metric::Enumerated type, giving rise to distance function $δ$. See distanceSqr. By default, the Fisher metric is adopted.

See also: laplacian, laplacianEigenMaps, spectralEmbedding.

Examples

using PosDefManifold
# Generate a set of 4 random 10x10 SPD matrices
℘=randP(10, 4)
# Compute the squared inter-distance matrix according to the log Euclidean metric.
# This is much faster as compared to the Fisher metric and in general
# it is a good approximation.
Δ²=distanceSqrMat(℘, logEuclidean)

source

PosDefManifold.distanceMatrix — Function.

distanceMatrix(℘, metric::Metric=Fisher)

alias: distanceMat

Given a 1d array ℘ of $k$ positive definite matrices ${P_1,...,P_k}$, create the $k⋅k$ real Hermitian matrix comprising elements $δ(P_i, P_j)\textrm{, for all }i≠j$.

This is the matrix of all inter-distances (zero on diagonal), using the specified metric, of type Metric::Enumerated type, giving rise to distance $δ$. See distance. By default, the Fisher metric is adopted.

The elements of this matrix are the square root of distanceSqrMat.

Examples

using PosDefManifold
# Generate a set of 4 random 10x10 SPD matrices
℘=randP(10, 4)
Δ=distanceMatrix(℘)

source

PosDefManifold.laplacian — Function.

laplacian(Δ²)

Given a matrix of squared inter-distances $Δ^2$, computed for examples by function distanceSqrMat, return the normalized Laplacian.

First, a Gaussian radial basis functions is applied to all elements of $Δ^2$, such as

$W_{ij} = exp(\frac{\displaystyle{-Δ^2_{ij}}}{\displaystyle{ε}})$,

where $ε$ is the Gaussian scale parameter chosen automatically as the median of the elements $Δ^2_{ij}$.

Finally, the normalized Laplacian is defined as

$Ω = D^{-1/2}WD^{-1/2}$,

where $D$ is the diagonal matrix holding on the main diagonal the sum of the rows (or columns) of $W$.

Nota Bene

The normalized Laplacian as here defined can be requested for any input matrix of squared inter-distances, for example, those obtained on scalars or on vectors using appropriate metrics.

See also: distanceSqrMat, laplacianEigenMaps, spectralEmbedding.

Examples

using PosDefManifold
# Generate a set of 4 random 10x10 SPD matrices
℘=randP(10, 4)
Δ²=distanceSqrMat(℘)
Ω=laplacian(Δ²) # or, equivalently, Ω=RΩ(Δ)

source

PosDefManifold.laplacianEigenMaps — Function.

laplacianEigenMaps(Ω, q::Int; <tol=1e-9, maxiter=300, ⍰=false>)

alias: laplacianEM

Given a normalized Laplacian $Ω$ (see laplacian ) return the eigen maps in $q$ dimensions, i.e., the $q$ eigenvectors of the normalized Laplacian associated with the largest $q$ eigenvalues, excluding the first (which is always equal to 1.0).

The eigenvectors of the normalized Laplacian are computed by the power iterations+modified Gram-Schmidt method, allowing calling this function even for big Laplacian matrices.

Return the 4-tuple $(Λ, U, iterations, convergence)$, where:

$Λ$ is a $q⋅q$ diagonal matrix holding on diagonal the eigenvalues corresponding to the $q$ dimensions of the Laplacian eigen maps;

$U$ holds in columns the eigen maps, that is, the $q$ eigenvectors

$iterations$ is the number of iterations executed by the power method;

$convergence$ is the convergence attained by the power method;

The eigenvectors of $U$ holds the coordinates of the points in the embedded space. For examples of applications see Ridrigues et al. (2018) 🎓.

Nota Bene

The maximum value of $q$ that can be requested is $n-1$, where $n$ is the size of the Laplacian. In general, $q=2$ or $q=3$ is requested.

Arguments: (Ω, q; <tol=1e-9, maxiter=300, ⍰=false>):

$Ω$ is a normalized Laplacian obtained by the laplacian function;

$q$ is the dimension of the Laplacian eigen maps;

The following are <optional keyword arguments> for the power method iterative algorithm:

tol is the tolerance for convergence;

maxiter is the maximum number of iterations allowed;

if ⍰ is true, the convergence at all iterations will be printed.

See also: distanceSqrMat, laplacian, spectralEmbedding.

Examples

using PosDefManifold
# Generate a set of 4 random 10x10 SPD matrices
℘=randP(10, 4)
evalues, maps, iterations, convergence=laplacianEM(Ω, 2)
evalues, maps, iterations, convergence=laplacianEM(Ω, 2, maxiter=500)
evalues, maps, iterations, convergence=laplacianEM(Ω, 2, ⍰=true)

source

PosDefManifold.spectralEmbedding — Function.

spectralEmbedding(℘, q::Int, metric::Metric=Fisher;
                    <tol=1e-9, maxiter=300, ⍰=false>)

alias: Rse

Given a 1d array ℘ of $k$ positive definite matrices ${P_1,...,P_k}$, compute its eigen maps in $q$ dimensions.

This function runs one after the other the functions:

distanceSqrMat (compute the squared inter-distance matrix),

laplacian (compute the normalized Laplacian),

laplacianEigenMaps (get the eigen maps).

Return the 4-tuple (Λ, U, iterations, convergence), where:

$Λ$ is a $q⋅q$ diagonal matrix holding on diagonal the eigenvalues corresponding to the $q$ dimensions of the Laplacian eigen maps;

$U$ holds in columns the $q$ eigenvectors, i.e., the $q$ coordinates of the points in the embedded space.

$iterations$ is the number of iterations executed by the power method;

$convergence$ is the convergence attained by the power method;

Arguments (℘, q, metric, <tol=1e-9, maxiter=300, ⍰=false>):

℘ is a 1d array of $k$ positive matrices;

$q$ is the dimension of the Laplacian eigen maps;

metric is a metric of type Metric::Enumerated type, used for computing the inter-distances. By default, the Fisher metric is adopted.

The following are <optional keyword arguments> for the power method iterative algorithm:

tol is the tolerance for convergence of the power method;

maxiter is the maximum number of iterations allowed for the power method;

if ⍰ is true the convergence at all iterations will be printed.

See also: distanceSqrMat, laplacian, laplacianEigenMaps.

Examples

using PosDefManifold
# Generate a set of 4 random 10x10 SPD matrices
℘=randP(10, 4)
evalues, maps, iterations, convergence=spectralEmbedding(℘, 2)
evalues, maps, iterations, convergence=spectralEmbedding(℘, 2, ⍰=true)

source

Means

PosDefManifold.generalizedMean — Function.

generalizedMean(℘, p::Real; <w::Vector=[], ✓w::Bool=true>)

Given a 1d array ℘ of $k$ positive definite matrices${P_1,...,P_k}$ and optional non-negative real weights vector $w={w_1,...,w_k}$, return the weighted generalized mean $G$ with real parameter $p$, that is,

$G=\big(\sum_{i=1}^{k}w_iP_i^p\big)^{1/p}$.

If you don't pass a weight vector with <optional keyword argument> $w$, return the unweighted generalized mean.

$G=\big(\sum_{i=1}^{k}P_i^p\big)^{1/p}$.

If <optional keword argument> ✓w=true (default), the weights are normalized so as to sum up to 1, otherwise they are used as they are passed. This option is provided to allow calling this function repeatedly without normalizing the weights each time.

The following special cases for parameter $p$ are noteworthy:

For $p=\frac{1}{2}$ the generalized mean is the modified Bhattacharyya mean.

For $p=1$ the generalized mean is the Euclidean mean.

For $p=-1$ the generalized mean is the inverse Euclidean mean.

For $p=0$ the generalized mean is the log Euclidean mean, which is the Fisher mean when matrices in ℘ all pair-wise commute.

Notice that when matrices in ℘ all pair-wise commute, the generalized means coincide with the power means for any $p∈[-1, 1]$ and for $p=0.5$ it coincides also with the Wasserstein mean (see wasMean). For this reason the generalized means are used as default initialization of both the powerMean and wasMean algorithm.

See also: powerMean

Examples

using LinearAlgebra, Statistics, PosDefManifold
# Generate a set of 4 random 3x3 SPD matrices
℘=randP(3, 4)

# weights vector, does not need to be normalized
weights=[1, 2, 3, 1]

# unweighted mean
G = generalizedMean(℘, 0.25)

# weighted mean
G = generalizedMean(℘, 0.5; w=weights)

# with weights previously normalized we can set ✓w=false
weights=weights./mean(weights)
G = generalizedMean(℘, 0.5; w=weights, ✓w=false)

source

PosDefManifold.logdet0Mean — Function.

logdet0Mean(℘; <w::Vector=[], ✓w::Bool=true, init=nothing,
                 tol=1e-9, ⍰=false>)

Given a 1d array $℘$ of $k$ positive definite matrices ${P_1,...,P_k}$ and optional non-negative real weights vector $w={w_1,...,w_k}$, return the 3-tuple $(G, iter, conv)$, where $G$ is the mean according to the logdet zero metric and $iter$, $conv$ are the number of iterations and convergence attained by the algorithm. Mean $G$ is the unique positive definite matrix satisfying

$\sum_{i=1}^{k}w_i\big(\frac{1}{2}P_i+\frac{1}{2}G\big)^{-1}=G^{-1}$.

For estimating it, this function implements the fixed-point iteration algorithm suggested by (Moakher, 2012, p315)🎓, yielding iterations

$G ← \frac{1}{2}\big(\sum_{i=1}^{k}w_i(P_i+G)^{-1}\big)^{-1}$.

If you don't pass a weight vector with <optional keyword argument> $w$, return the unweighted logdet zero mean.

If <optional keword argument> ✓w=true (default), the weights are normalized so as to sum up to 1, otherwise they are used as they are passed and should be already normalized. This option is provided to allow calling this function repeatedly without normalizing the same weights vector each time.

The following are more <optional keyword arguments>:

init is a matrix to be used as initialization for the mean. If no matrix is provided, the log Euclidean mean will be used;

tol is the tolerance for the convergence. The smaller this number (it must be positive) the closer the algorithm gets to the saddle point;

if ⍰ is true, the convergence attained at each iteration is printed.

Nota Bene

In normal circumstances this algorithm converges monothonically. If the algorithm diverges a warning is printed indicating the iteration when this happened and the algorithm is interrupted.

See: logdet zero metric, modified Bhattacharyya mean.

Examples

using LinearAlgebra, PosDefManifold
# Generate a set of 4 random 3x3 SPD matrices
℘=randP(3, 4)

# unweighted mean
G, iter, conv = logdet0Mean(℘)

# weights vector, does not need to be normalized
weights=[1, 2, 3, 1]

# weighted mean
G, iter, conv = logdet0Mean(℘, w=weights)

# print the convergence at all iterations
G, iter, conv = logdet0Mean(℘, w=weights, ⍰=true)

# now suppose ℘ has changed a bit, initialize with G to hasten convergence
℘[1]=ℍ(℘[1]+(randP(3)/100))
G, iter, conv = logdet0Mean(℘, w=weights, ✓w=false, ⍰=true, init=G)

source

PosDefManifold.wasMean — Function.

wasMean(℘; <w::Vector=[], ✓w::Bool=true, init=nothing,
             tol=1e-9, ⍰=false>)

Given a 1d array ℘ of $k$ positive definite matrices ${P_1,...,P_k}$ and optional non-negative real weights vector $w={w_1,...,w_k}$, return the 3-tuple $(G, iter, conv)$, where $G$ is the mean according to the Wasserstein metric and $iter$, $conv$ are the number of iterations and convergence attained by the algorithm. Mean $G$ is the unique positive definite matrix satisfying

$G=\sum_{i=1}^{k}w_i\big( G^{1/2} P_i G^{1/2}\big)^{1/2}$.

For estimating it, this function implements the fixed-point iterative algorithm proposed by (Álvarez-Esteban et al., 2016)🎓:

$G ← G^{-1/2}\big(\sum_{i=1}^{k} w_i(G^{1/2}P_i G^{1/2})^{1/2}\big)^2 G^{-1/2}$.

If you don't pass a weight vector with <optional keyword argument> $w$, return the unweighted Wassertein mean.

If <optional keword argument> ✓w=true (default), the weights are normalized so as to sum up to 1, otherwise they are used as they are passed and Metric::Enumerated type be already normalized. This option is provided to allow calling this function repeatedly without normalizing the same weights vector each time.

The following are more <optional keyword arguments>:

init is a matrix to be used as initialization for the mean. If no matrix is provided, the instance of generalized means with $p=0.5$ will be used;

tol is the tolerance for the convergence. The smaller this number (it must be positive) the closer the algorithm gets to the true solution;

if ⍰ is true, the convergence attained at each iteration is printed.

Nota Bene

In normal circumstances this algorithm converges monothonically. If the algorithm diverges a warning is printed indicating the iteration when this happened and the algorithm is interrupted.

See: Wasserstein metric

Examples

using LinearAlgebra, PosDefManifold
# Generate a set of 4 random 3x3 SPD matrices
℘=randP(3, 4)

# unweighted mean
G, iter, conv = wasMean(℘)

# weights vector, does not need to be normalized
weights=[1, 2, 3, 1]

# weighted mean
G, iter, conv = wasMean(℘, w=weights)

# print the convergence at all iterations
G, iter, conv = wasMean(℘, w=weights, ⍰=true)

# now suppose ℘ has changed a bit, initialize with G to hasten convergence
℘[1]=ℍ(℘[1]+(randP(3)/100))
G, iter, conv = wasMean(℘, w=weights, ⍰=true, init=G)

source

PosDefManifold.powerMean — Function.

powerMean(℘, p::Real; <w::Vector=[], ✓w::Bool=true, init=nothing,
                        tol=1e-9, ⍰=false>)

Given a 1d array ℘ of $k$ positive definite matrices ${P_1,...,P_k}$, an optional non-negative real weights vector $w={w_1,...,w_k}$ and a real parameter p $\in[-1, 1]$, return the 3-tuple $(G, iter, conv)$, where $G$ is Lim and Palfia (2012)'s power means of order $p$ and $iter$, $conv$ are the number of iterations and convergence attained by the algorithm, respectively. Mean $G$ is the unique positive definite matrix satisfying

$G=\sum_{i=1}^{k}(w_iG\textrm{#}_pP_i)$,

where $G\textrm{#}_pP_i$ is the Fisher geodesic equation. In particular:

with $p=-1$ this is the harmonic mean (see the inverse Euclidean)

with $p=+1$ this is the arithmetic mean (see the Euclidean)

at the limit of $p$ evaluated at zero from both side this is the geometric mean (see the Fisher metric).

For estimating power means for $p\in(-1, 1)$, this function implements the fixed-point iterative algorithm of (Congedo et al., 2017b)🎓. For $p=0$ (geometric mean) this algorithm is run two times with a small positive and negative value of $p$ and the geometric mean of the two resulting means is returned, as suggested in (Congedo et al., 2017b)🎓. This way of estimating the geometric mean of a set of matrices is faster as compared to the usual gradient descent algorithm.

If you don't pass a weight vector with <optional keyword argument> $w$, return the unweighted power mean.

If <optional keword argument> ✓w=true (default), the weights are normalized so as to sum up to 1, otherwise they are used as they are passed and Metric::Enumerated type be already normalized. This option is provided to allow calling this function repeatedly without normalizing the same weights vector each time.

The following are more <optional keyword arguments>:

init is a matrix to be used as initialization for the mean. If no matrix is provided, the instance of generalized means with parameter $p$ will be used.

tol is the tolerance for the convergence. The smaller this number (it must be positive) the closer the algorithm gets to the true solution;

if ⍰ is true, the convergence attained at each iteration is printed.

Nota Bene

In normal circumstances this algorithm converges monothonically. If the algorithm diverges a warning is printed indicating the iteration when this happened and the algorithm is interrupted.

See: modified Bhattacharyya mean

Examples

using LinearAlgebra, PosDefManifold
# Generate a set of 4 random 3x3 SPD matrices
℘=randP(3, 4)

# unweighted mean
G, iter, conv = powerMean(℘, 0.5)

# weights vector, does not need to be normalized
weights=[1, 2, 3, 1]

# weighted mean
G, iter, conv = powerMean(℘, 0.5, w=weights)

# print the convergence at all iterations
G, iter, conv = powerMean(℘, 0.5, w=weights, ⍰=true)

# now suppose ℘ has changed a bit, initialize with G to hasten convergence
℘[1]=ℍ(℘[1]+(randP(3)/100))
G, iter, conv = powerMean(℘, 0.5, w=weights, ⍰=true, init=G)

source

PosDefManifold.meanP — Function.

(1) meanP(P::ℍ, Q::ℍ, metric::Metric=Fisher)
(2) meanP(℘, metric::Metric=Fisher; <w::Vector=[], ✓w::Bool=true>)

(1) Mean of two positive definite matrices, passed in arbitrary order as arguments $P$ and $Q$, using the specified metric of type Metric::Enumerated type. By defult the Fisher metric is used. The order is arbitrary as all metrics implemented in PosDefManifold are symmetric. This is the midpoint of the geodesic. For the weighted mean of two positive definite matrices use instead the geodesic function. $P$ and $Q$ must be flagged as Hermitian. See typecasting matrices.

(2) Fréchet mean of an 1d array $℘$ of $k$ positive definite matrices${P_1,...,P_k}$, with optional non-negative real weights $w={w_1,...,w_k}$ using the specified metricas in (1).

If you don't pass a weight vector with <optional keyword argument> $w$, return the unweighted mean.

If <optional keword argument> ✓w=true (default), the weights are normalized so as to sum up to 1, otherwise they are used as they are passed and should be already normalized. This option is provided to allow calling this function repeatedly without normalizing the same weights vector each time.

Math

The Fréchet mean of a set of $k$ matrices ${P_1, P_2,..., P_k}$ weighted by ${w_1, w_2,..., w_k}:\sum_{i=1}^{k}w_i=1$ for the supported metrics are, for those with closed form expression:

Metric	weighted Fréchet mean

Euclidean	$\sum_{i=1}^{k}w_i P_i$

invEuclidean	$\big(\sum_{i=1}^{k}w_i P_i^{-1}\big)^{-1}$

ChoEuclidean	$TT^*$, where $T=bL_P + aL_Q$

logEuclidean	$\textrm{exp}\big(\sum_{i=1}^{k}w_i\hspace{1pt} \textrm{log}P_i \big)$

logCholesky	$TT^*$, where $T=\sum_{i=1}^{k}(w_kS_k)+\sum_{i=1}^{k}(w_k\textrm{log}D_k)$

Jeffrey	$A^{1/2}\big(A^{-1/2}HA^{-1/2}\big)^{1/2}A^{1/2}$

and for those that verify an equation:

Metric	equation verified by the weighted Fréchet mean

Fisher	$\sum_{i=1}^{k}w_i\textrm{log}\big(G^{-1/2} P_k G^{-1/2}\big)=0.$

logdet0	$\sum_{i=1}^{k}w_i\big(\frac{1}{2}P_i+\frac{1}{2}G\big)^{-1}=G^{-1}$

VonNeumann	N.A.

Wasserstein	$G=\sum_{i=1}^{k}w_i\big( G^{1/2} P_i G^{1/2}\big)^{1/2}$

legend: $L_X$, $S_X$ and $D_X$ are the Cholesky lower triangle of $X$, its strictly lower triangular part and diagonal part, respectively (hence, $S_X+D_X=L_X$, $L_XL_X^*=X$). $A$ and $H$ are the weighted arithmetic and weighted harmonic mean, respectively.

See: geodesic, mean.

Examples

using LinearAlgebra, Statistics, PosDefManifold
# Generate 2 random 3x3 SPD matrices
P=randP(3)
Q=randP(3)
M=meanP(P, Q, logdet0) # (1)
M=meanP(P, Q) # (1), uses Fisher metric

# Generate a set of 4 random 3x3 SPD matrices
℘=randP(3, 4)
# weights vector, does not need to be normalized
weights=[1, 2, 3, 1]
M=meanP(℘, Euclidean, w=weights) # (2) weighted Euclidean mean
M=meanP(℘, Wasserstein)  # (2) unweighted Wassertein mean

source

Tangent Space operations

PosDefManifold.logMap — Function.

logMap(P::ℍ, G::ℍ, metric::Metric=Fisher)

Logaritmic Map: map a positive definite matrix $P$ from the SPD or Hermitian manifold into the tangent space at base-point $G$ using the Fisher metric.

$P$ and $G$ must be flagged as Hermitian. See typecasting matrices.

The map is defined as

$Log_G(P)=S=G^{1/2}\textrm{log}\big(G^{-1/2}PG^{-1/2}\big)G^{1/2}$.

The result is an Hermitian matrix. The inverse operation is expMap.

Arguments (P, G, metric):

$P$ is the positive definite matrix to be projected onto the tangent space,

$G$ is the tangent space base point,

metric is a metric of type Metric::Enumerated type.

Currently only the Fisher metric is supported for tangent space operations.

See also: vecP

Examples

using PosDefManifold
P=randP(3)
Q=randP(3)
G=meanP(P, Q)
# projecting P at the base point given by the geometric mean of P and Q
S=logMap(P, G)

source

PosDefManifold.expMap — Function.

expMap(S::ℍ, G::ℍ, metric::Metric=Fisher)

Exponential Map: map an Hermitian matrix $S$ from the tangent space at base point $G$ into the SPD or Hermitian manifold (using the Fisher metric).

$S$ and $G$ must be flagged as Hermitian. See typecasting matrices.

The map is defined as

$Exp_G(S)=P=G^{1/2}\textrm{exp}\big(G^{-1/2}SG^{-1/2}\big)G^{1/2}$.

The result is a positive definite matrix. The inverse operation is logMap.

Arguments (S, G, metric):

$S$ is a Hermitian matrix, real or complex, to be projected on the SPD or Hermitian manifold,

$G$ is the tangent space base point,

metric is a metric of type Metric::Enumerated type.

Currently only the Fisher metric is supported for tangent space operations.

Examples

using PosDefManifold, LinearAlgebra
P=randP(3)
Q=randP(3)
G=meanP(P, Q, Fisher)
# projecting P on the tangent space at the Fisher mean base point G
S=logMap(P, G)
# adding the identity in the tangent space and reprojecting back onto the manifold
H=expMap(ℍ(S+I), G)

source

PosDefManifold.vecP — Function.

vecP(S::ℍ)

Vectorize a tangent vector (matrix) $S$ (i.e., an Hermitian matrix): mat -> vec.

It gives weight $1$ to diagonal elements and √2 to off-diagonal elements (Barachant et al., 2012)🎓.

The result is a vector holding $n(n+1)/2$ elements, where $n$ is the size of $S$.

$S$ must be flagged as Hermitian. See typecasting matrices.

The inverse operation is provided by matP.

Examples

using PosDefManifold
P=randP(3)
Q=randP(3)
G=meanP(P, Q, Fisher)
# projecting P at the base point given by the geometric mean of P and Q
S=logMap(P, G)
# vectorize S
ς=vecP(S)

source

PosDefManifold.matP — Function.

matP(ς::Vector)

Matrizize a tangent vector (vector) ς : vec -> mat.

This is the function reversing the vecP function, thus the weighting applied therein is reversed as well.

If $ς=vecP(S)$ and $S$ is a $n⋅n$ Hermitian matrix, $ς$ is a tangent vector of size $n(n+1)/2$. The result of calling $matP(ς)$ is then $n⋅n$ matrix $S$.

P.S.: This function needs to be rewritten more efficiently

Examples

using PosDefManifold
P=randP(3)
Q=randP(3)
G=meanP(P, Q, Fisher)
# projecting P at onto the tangent space at the Fisher mean base point
S=logMap(P, G)
# vectorize S
ς=vecP(S)
# Rotate the vector by an orthogonal matrix
n=Int(size(S, 1)*(size(S, 1)+1)/2)
U=randP(n)
v=U*ς
# Get the point in the tangent space
S=matP(v)

source

Procrustes problems

PosDefManifold.procrustes — Function.

procrustes(P::ℍ, Q::ℍ, extremum="min")

Given two positive definite matrices $P$ and $Q$, return by default the solution of problem

$\textrm{argmin}_Uδ(P,U^*QU)$,

where $U$ varies over the set of unitary matrices $𝐔$ and $δ(.,.)$ is a distance or divergence function. $U^*QU$ is named in physics the unitary orbit of $Q$.

If the argument 'extremum' is passed as "max", it returns instead the solution of

$\textrm{argmax}_Uδ(P,U^*QU)$.

$P$ and $Q$ must be flagged as Hermitian. See typecasting matrices.

As it has been shown in Bhatia and Congedo (2019)🎓, using each of the Fisher, logdet zero, Wasserstein and the Kullback-Leibler divergence (see logdet α), the best approximant to $P$ from the unitary orbit of $Q$ commutes with $P$ and, surprisingly, has the same closed-form expression, namely

$U_Q^↓U_P^{↓*}$ for the argmin and $U_Q^↑U_P^{↓*}$ for the argmax,

where $U^↓$ denotes the eigenvector matrix of the subscript argument with eigenvectors in columns sorted by decreasing order of corresponding eigenvalues and $U^↑$ denotes the eigenvector matrix of the subscript argument with eigenvectors in columns sorted by increasing order of corresponding eigenvalues.

The same solutions are known since a long time also by solving the extremal problem here above using the Euclidean metric (Umeyama, 1988).

Examples

using PosDefManifold
P=randP(3)
Q=randP(3)
# argmin problem
U=procrustes(P, Q)
# argmax problem
V=procrustes(P, Q, "max")

source